Correct start coordinate untemplated nts on -
Reads mapping in reverse orientation that were cleaned up for untemplated nts at their 3' end got a correction of their
genomic end position instead of the begin position. This has been corrected. Testing on prediction results due to this change
are still to be performed. An effect is not to be expected if the proportion of reads that are used to predict novel miRNAs do not have untemplated nts at their 3'end and additionally are traced back to the reverse complementary genome sequence.